switch int x and y to double
# add methot to get random double location
# location is now stored as double type

+++ SmartGun/MessagesLibrary/ShotMessage.cs
@@ -5,7 +5,7 @@ namespace MessagesLibrary
     [DataContract]
     public class ShotMessage : Message
     {
-        public ShotMessage(int x, int y, string username) : base(x, y, username)
+        public ShotMessage(double x, double y, string username) : base(x, y, username)
         {
             MessageType = MessageType.Shot;
         }
