+++ SmartGun/MessagesLibrary/UpdateMessage.cs
@@ -6,7 +6,7 @@ namespace MessagesLibrary
     public class UpdateMessage:Message
     {
 
-        public UpdateMessage(int x, int y, string username) : base(x, y, username)
+        public UpdateMessage(double x, double y, string username) : base(x, y, username)
         {
             MessageType = MessageType.Update;
         }
